:sparkles: feat : 기프티콘 OCR, S3 업로드, 삭제 기능 추가 (임시 완성) OCR 텍스트 핸들링 로직 미완성

+++ backend/src/main/java/com/ssafy/common/util/AmazonS3Util.java
@@ -1,10 +1,8 @@
 package com.ssafy.common.util;
 
 import com.amazonaws.services.s3.AmazonS3Client;
-import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.google.api.client.util.Value;
 import com.ssafy.common.exception.CustomException;
 import com.ssafy.common.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
@@ -14,11 +12,9 @@ import org.springframework.stereotype.Component;
 import org.springframework.util.ObjectUtils;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.Optional;
 
 @RequiredArgsConstructor
 @Component
@@ -33,8 +29,8 @@ public class AmazonS3Util {
     public String upload(MultipartFile multipartFile, String userAccessToken) throws IOException { // 객체 업로드
 
         String originalFileName = multipartFile.getOriginalFilename();
-        String ext = originalFileName.substring(originalFileName.lastIndexOf(".") + 1);
-        String contentType = multipartFile.getContentType(); // 파일의 확장자 추출
+        String ext = originalFileName.substring(originalFileName.lastIndexOf(".") + 1); // 파일 확장자(png)
+        String contentType = multipartFile.getContentType(); // 파일 확장자(image/png)
 
         long size = multipartFile.getSize(); // 파일 크기
 
@@ -50,50 +46,31 @@ public class AmazonS3Util {
             throw new CustomException(ErrorCode.INVALID_REQUEST); // 나중에 수정 필요
         }
 
-        String url = null;
-
-//        File uploadFile = convert(multipleFile)
-//                .orElseThrow(() -> new IllegalArgumentException("MultipartFile -> File로 전환이 실패했습니다."));
-
-        // 파일명 중복을 피하기위해 날짜 추가
+        // 파일명 중복을 피하기위해 유저 토큰 + 타임스탬프 추가
         String formatDate = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMddhhmmssSSS"));
-
-        // String fileName = dirName + formatDate + uploadFile.getName();
         String fileName = userAccessToken + formatDate + "." + ext; // 최종 파일명
 
+        // 이미 똑같은 파일명이 존재하는 경우
+        boolean isExistObject = amazonS3Client.doesObjectExist(bucket, fileName);
+        if(isExistObject) throw new CustomException(ErrorCode.INVALID_REQUEST); // 나중에 수정 필요
+
+        // S3로 업로드
         amazonS3Client.putObject(
                 new PutObjectRequest(bucket, fileName, multipartFile.getInputStream(), objectMetaData)
         );
 
-        // put - S3로 업로드
-        String uploadImageUrl = amazonS3Client.getUrl(bucket, fileName).toString();
-
-//        // 로컬 파일 삭제
-//        removeFile(uploadFile);
-
-        url = uploadImageUrl;
-        return url;
+        return fileName;
     }
 
-    private Optional<File> convert(MultipartFile file) throws IOException { // multipartFile -> File
-        File convertFile = new File(file.getOriginalFilename());
-        file.transferTo(convertFile);
-        return Optional.of(convertFile);
-    }
-
-    private String putS3(File uploadFile, String fileName) { // S3로 업로드
-        amazonS3Client.putObject(new PutObjectRequest(bucket, fileName, uploadFile).withCannedAcl(CannedAccessControlList.BucketOwnerRead));
-        return amazonS3Client.getUrl(bucket, fileName).toString();
-    }
-
-    private void removeFile(File targetFile) { // 로컬파일 삭제
-        if (targetFile.exists()) {
-            if (targetFile.delete()) {
-                log.info("파일이 삭제되었습니다.");
-            } else {
-                log.info("파일이 삭제되지 못했습니다.");
-            }
+    public boolean delete(String filePath) {
+        boolean isExistObject = amazonS3Client.doesObjectExist(bucket, filePath);
+        if(isExistObject) {
+            amazonS3Client.deleteObject(bucket, filePath);
+            return true;
+        } else {
+            return false;
         }
+
     }
 
 }
